backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.3
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
supabase-py==2.3.5 